api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
numpy
scikit-learn
joblib
python-dotenv
